Model/src/main/java/ro/ubb/pm/model/RememberMeToken.java
package ro.ubb.pm.model;

import javax.persistence.Column;
import javax.persistence.Table;
import javax.validation.constraints.NotNull;

@javax.persistence.Entity
@Table(name = "remember_me_tokens")
public class RememberMeToken extends Entity {

    @NotNull
    @Column(name = "email")
    private String email;

    @NotNull
    @Column(name = "password")
    private String password;

    public RememberMeToken() {}

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }

    public String getPassword() {
        return password;
    }

    public void setPassword(String password) {
        this.password = password;
    }
}
